# [mono] Don't use recursion in mono_ssa_rename_vars